Remove dependency on hive-exec and instead use hive-exec:core
The hive-exec library bundles all its dependencies and this creates conflicts
with Connect's libraries (for example, avro-1.8.2 in Connect conflicts with the
avro(-1.7.7) classes bundled in hive-exec). We also need to add commons-lang3
and kryo and avro-mapred here, since they are not pulled in by hive-exec's pom
but required at runtime.

Signed-off-by: Arjun Satish <arjun@confluent.io>