CC-5963: Improved ability to handle schema evolution with different names
Added additional checks to `StorageSchemaCompatibility.shouldChangeSchema(…)` so that this method considers not only the versions but also considers a schema as having changed when the name is changed, the types are different (but not projectable), or the parameters are changed. This makes this logic more compatible with the `project(…)` methods on the same enumeration used to project a record into the schema used in a file with lots of records.

Thus, connectors that use this will effectively rotate whenever such schema evolution occurs and projection (using Connect’s SchemaProjector) would not work successfully. Most users don’t have this kind of schema evolution, and thus would see no change in behavior. And users that do have this kind of schema evolution likely do so infrequently.

This change is still compatible with Schema Registry and Avro’s schema compatiblity behavior. In particular, Avro allows schemas to have different names and namespaces as long as the field requirements are still satisfied. So when a user is using SR and backward, forward, or full compatbility, SR and the Avro serializer will not prevent writing a record with an evolved schema. With this change, the records on a topic consumed by a connector using this class will consider the evolved schema as “changed”, and the connector will rotate its file by closing it, creating a new one, and writing the record with the new schema to the new file.

Quite a few unit tests were added to verify the StorageSchemaCompatibility behaviors, including the new behavior.